+++ src/main/java/bowling/domain/frame/FinalFrame.java
@@ -3,8 +3,7 @@ package bowling.domain.frame;
 import java.util.Collections;
 import java.util.List;
 
-import bowling.domain.KnockedPins;
-import bowling.domain.exception.InvalidMethodCallException;
+import bowling.domain.common.KnockedPins;
 import bowling.domain.frame.exception.InvalidProgressFrameException;
 import bowling.domain.pitch.Pitch;
