+++ src/main/java/bowling/domain/frame/Frame.java
@@ -2,7 +2,7 @@ package bowling.domain.frame;
 
 import java.util.List;
 
-import bowling.domain.KnockedPins;
+import bowling.domain.common.KnockedPins;
 import bowling.domain.pitch.Pitch;
 
 public interface Frame {
