+++ src/test/java/bowling/domain/common/FalledPinsTest.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.common;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -8,7 +8,7 @@ import org.junit.jupiter.params.provider.CsvSource;
 import bowling.domain.exception.InvalidPinCountException;
 
 @DisplayName("쓰러진 핀")
-class KnockedPinsTest {
+class FalledPinsTest {
 
 	@DisplayName("[실패] 생성 - 유효하지 않은 핀 개수")
 	@ParameterizedTest
@@ -19,7 +19,7 @@ class KnockedPinsTest {
 		// given
 
 		// when
-		Assertions.assertThrows(InvalidPinCountException.class, () -> KnockedPins.of(pinCount));
+		Assertions.assertThrows(InvalidPinCountException.class, () -> FalledPins.of(pinCount));
 
 		// then
 	}
