+++ src/test/java/bowling/domain/frame/NormalFrameTest.java
@@ -9,7 +9,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
 import bowling.domain.frame.exception.InvalidProgressFrameException;
-import bowling.domain.frame.exception.PitchesFullException;
 
 @DisplayName("일반 프레임")
 class NormalFrameTest {
