+++ src/test/java/bowling/domain/pitch/StrikePitchTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import bowling.domain.KnockedPins;
+import bowling.domain.common.KnockedPins;
 
 @DisplayName("투구 - 스트라이크")
 class StrikePitchTest {
